+++ core/src/services/routing.js
@@ -234,6 +234,33 @@ class RoutingClass {
           ? pathData.navigationPath[pathData.navigationPath.length - 1]
           : null;
 
+      let tabNavInherited = false;
+      let cnode = currentNode;
+      while (cnode) {
+        if (cnode.tabNav === true) {
+          tabNavInherited = true;
+          break;
+        } else if (cnode.tabNav === false) {
+          tabNavInherited = false;
+          break;
+        }
+        cnode = cnode.parent;
+      }
+
+      let cNode2 = currentNode;
+      let hideSideNavInherited = nodeObject.hideSideNav;
+      while (cNode2) {
+        if (cNode2.tabNav && cNode2.hideSideNav === true) {
+          hideSideNavInherited = true;
+          break;
+        }
+        if (cNode2.hideSideNav === false) {
+          hideSideNavInherited = false;
+          break;
+        }
+        cNode2 = cNode2.parent;
+      }
+
       const newNodeData = {
         hideNav,
         viewUrl,
@@ -247,8 +274,9 @@ class RoutingClass {
           pathData.pathParams
         ),
         pathParams: pathData.pathParams,
-        hideSideNav: nodeObject.hideSideNav || false,
-        isolateView: nodeObject.isolateView || false
+        hideSideNav: hideSideNavInherited || false,
+        isolateView: nodeObject.isolateView || false,
+        tabNav: tabNavInherited
       };
 
       const previousCompData = component.get();
@@ -264,6 +292,18 @@ class RoutingClass {
         })
       );
 
+      let iContainer = document.getElementsByClassName('iframeContainer')[0];
+      if (iContainer) {
+        if (tabNavInherited) {
+          //document.body.classList.add('lui-simpleSlideInNav');
+          iContainer.classList.add('iframeContainerTabNav');
+        } else {
+          if (iContainer.classList.contains('iframeContainerTabNav')) {
+            iContainer.classList.remove('iframeContainerTabNav');
+          }
+        }
+      }
+
       Iframe.navigateIframe(config, component, iframeElement);
     } catch (err) {
       console.info('Could not handle route change', err);
@@ -272,7 +312,7 @@ class RoutingClass {
 
   handleRouteClick(node, component) {
     const componentData = component.get();
-    if (node.externalLink?.url) {
+    if (node.externalLink && node.externalLink.url) {
       this.navigateToExternalLink(node.externalLink);
       // externalLinkUrl property is provided so there's no need to trigger routing mechanizm
     } else if (node.link) {
@@ -357,7 +397,7 @@ class RoutingClass {
   }
 
   navigateToLink(item) {
-    if (item.externalLink?.url) {
+    if (item.externalLink && item.externalLink.url) {
       this.navigateToExternalLink(item.externalLink);
     } else {
       this.navigateTo(item.link);