+++ core/webpack-common-rules.js
@@ -3,38 +3,40 @@ const sass = require('node-sass');
 
 module.exports = {
   svelte: {
-    test: /\.html$/,
+    test: /\.(svelte|html)$/,
     exclude: /node_modules/,
-    use: {
-      loader: 'svelte-loader',
-      options: {
-        emitCss: true,
-        name: 'Luigi',
-        preprocess: {
-          style: ({ content, attributes }) => {
-            if (attributes.type !== 'text/scss') return;
-            return new Promise((fulfil, reject) => {
-              sass.render(
-                {
-                  data: content,
-                  includePaths: ['src'],
-                  sourceMap: true,
-                  outFile: 'x' // this is necessary, but is ignored
-                },
-                (err, result) => {
-                  if (err) return reject(err);
+    use: [
+      {
+        loader: 'svelte-loader',
+        options: {
+          emitCss: true,
+          name: 'Luigi',
+          preprocess: {
+            style: ({ content, attributes }) => {
+              if (attributes.type !== 'text/scss') return;
+              return new Promise((fulfil, reject) => {
+                sass.render(
+                  {
+                    data: content,
+                    includePaths: ['src'],
+                    sourceMap: true,
+                    outFile: 'x' // this is necessary, but is ignored
+                  },
+                  (err, result) => {
+                    if (err) return reject(err);
 
-                  fulfil({
-                    code: result.css.toString(),
-                    map: result.map.toString()
-                  });
-                }
-              );
-            });
+                    fulfil({
+                      code: result.css.toString(),
+                      map: result.map.toString()
+                    });
+                  }
+                );
+              });
+            }
           }
         }
       }
-    }
+    ]
   },
   css: {
     test: /\.(sa|sc|c)ss$/,